Project/requirements.txt
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.0
blinker==1.6.2
cachelib==0.9.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
dnspython==2.3.0
email-validator==2.0.0.post1
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsons==1.6.3
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
pycountry==22.3.5
pycparser==2.21
PyMySQL==1.0.2
pytest==7.2.2
pytest-order==1.1.0
python-dotenv==1.0.0
soupsieve==2.4
SQLAlchemy==2.0.10
tomli==2.0.1
typing_extensions==4.5.0
typish==1.9.3
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
